refactor(core): shared store-column manifest and formalized normalizeRow seam
